requirements.txt
python-dotenv==1.0.0
openai>=1.3.0
google-cloud-documentai>=2.20.0
python-docx==0.8.11
Pillow==10.1.0